refactor(index): put defaults in own file, use merge inst. Object.assign
Object.assign does not deepmerge, merge from lodash does
